support for workspaces
On a workspaces based monorepo package-lock.json or yarn.lock
files could be at monorepo root instead of the package root.

This update will lookup for those files up the hierarchy if
they are not at the same same level as the package.json file.
Once one of them is found it stop searching.